backend/app/pyproject.toml
[tool.poetry]
name = "fastapi-graphql-starter"
version = "0.1.0"
description = ""
authors = ["Matthew Chung <matthewchung74@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
Jinja2 = "^2.11.3"
httpx = "^0.16.1"
graphene = "^2.1.8"
graphene-sqlalchemy = "^2.3.0"
async-exit-stack = "^1.0.1"
async_generator = "^1.10"
psycopg2-binary = "^2.8.6"
graphene-pydantic = "0.1.0"
bcrypt = "^3.2.0"
python-jose = "^3.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
asyncpg = "^0.22.0"
databases = {extras = ["postgresql"], version = "^0.4.1"}
alembic = "^1.5.4"
pytest-asyncio = "^0.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
